api/requirements.txt
fastapi
uvicorn
pytest
python-dotenv
python-multipart
